# docs/swagger.yaml
definitions:
...
  reqmodel.MovieRatingListResponse:
    properties:
      movie_id:
        type: string
      movie_rating_list:
        items:
          $ref: '#/definitions/sqlc.GetMovieRatingListRow'
        type: array
    type: object
...
  reqmodel.RatingCreateRequest:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
    type: object
...
  reqmodel.RatingUpdateRequest:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
    type: object
...
  reqmodel.UserRatingListResponse:
    properties:
      user_id:
        type: string
      user_rating_list:
        items:
          $ref: '#/definitions/sqlc.GetUserRatingListRow'
        type: array
    type: object
...
  sqlc.GetMovieRatingListRow:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
    type: object
...
  sqlc.GetUserRatingListRow:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
    type: object
...
  sqlc.Rating:
    properties:
      movie_id:
        type: string
      rating:
        type: integer
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Basic swagger for current api
  termsOfService: http://swagger.io/terms/
  title: movie_backend_go
  version: "1.0"
paths:
...
  /movie/{movie_id}/rating:
    get:
      consumes:
      - application/json
      description: Get users who rated movie
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.MovieRatingListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie rating list
      tags:
      - rating
      - movie
  /rating:
    delete:
      consumes:
      - application/json
      description: Delete certain rating
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Delete rating
      tags:
      - rating
      - admin
    patch:
      consumes:
      - application/json
      description: Update rating
      parameters:
      - description: Updated rating
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.RatingUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Rating'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Update rating
      tags:
      - rating
    post:
      consumes:
      - application/json
      description: Rate movie by user
      parameters:
      - description: Rate movie data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.RatingCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Rating'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Create rating
      tags:
      - rating
...
  /user/{user_id}/rating:
    get:
      consumes:
      - application/json
      description: Get user's rated movies
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.UserRatingListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user rating list
      tags:
      - rating
      - user
...
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: http://localhost:8080/auth/login
    type: oauth2
swagger: "2.0"